removed some warnings
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@3473 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/MSVehicle.h
@@ -668,7 +668,7 @@ protected:
                              SUMOTime currentTime);
 
     //compute accordant the distance, the Number of Infos that can be transmit
-    int numOfInfos(MSVehicle *veh1, MSVehicle* veh2);
+    size_t numOfInfos(MSVehicle *veh1, MSVehicle* veh2);
 
     //InfoCont getInfosToSend(void);
